requirements.txt
pytest
pytest-cov
pandas
boto3
psycopg2-binary